Add 'spent-virtually' concept, see below:
The idea is that sometimes small enough amount is used for some goal, so
there is no need for actual "withdrawal"- in such cases the user just
needs to "remember" to deposit less in the next deposit, and use the
amount for another goal.
So adding 'spent-virtually' box implements this idea- if spending is
marked 'virtual'- the amount is added to this box. Then, on the next
deposit, if not stated otherwise, the amount is added to 'free'